requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
appdirs=1.4.4=pyhd3eb1b0_0
attrs=22.1.0=py39h06a4308_0
beautifulsoup4=4.12.2=py39h06a4308_0
blas=1.0=mkl
boltons=23.0.0=py39h06a4308_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
certifi=2023.5.7=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
chardet=4.0.0=py39h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h06a4308_0
conda=23.5.0=py39h06a4308_0
conda-build=3.25.0=py39h06a4308_0
conda-index=0.2.3=py39h06a4308_0
conda-package-handling=2.1.0=py39h06a4308_0
conda-package-streaming=0.8.0=py39h06a4308_0
contourpy=1.0.5=py39hdb19cb5_0
coverage=7.2.2=py39h5eee18b_0
cryptography=39.0.1=py39h9ce1e76_2
cycler=0.11.0=pyhd3eb1b0_0
decorator=5.1.1=pyhd3eb1b0_0
exceptiongroup=1.0.4=py39h06a4308_0
filelock=3.9.0=py39h06a4308_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
future=0.18.3=py39h06a4308_0
giflib=5.2.1=h5eee18b_3
glob2=0.7=pyhd3eb1b0_0
greenlet=2.0.1=py39h6a678d5_0
icu=58.2=he6710b0_3
idna=3.4=py39h06a4308_0
importlib_resources=5.2.0=pyhd3eb1b0_1
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py39h06a4308_0
jpeg=9e=h5eee18b_1
jsonpatch=1.32=pyhd3eb1b0_0
jsonpointer=2.1=pyhd3eb1b0_0
kiwisolver=1.4.4=py39h6a678d5_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libarchive=3.6.2=h6ac8c49_2
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
liblief=0.12.3=h6a678d5_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.0=h6a678d5_2
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxml2=2.10.3=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lxml=4.9.2=py39h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib-base=3.7.1=py39h417a72b_1
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.6=py39h417a72b_1
mkl_random=1.2.2=py39h417a72b_1
more-itertools=8.12.0=pyhd3eb1b0_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
numpy=1.25.0=py39h5f9d8c6_0
numpy-base=1.25.0=py39hb5e798b_0
obspy=1.2.2=py39hce5d2b2_2
openssl=3.0.9=h7f8727e_0
packaging=23.0=py39h06a4308_0
patch=2.7.6=h7b6447c_1001
patchelf=0.17.2=h6a678d5_0
pillow=9.4.0=py39h6a678d5_0
pip=23.1.2=py39h06a4308_0
pkginfo=1.9.6=py39h06a4308_0
pluggy=1.0.0=py39h06a4308_1
pooch=1.4.0=pyhd3eb1b0_0
psutil=5.9.0=py39h5eee18b_0
py=1.11.0=pyhd3eb1b0_0
py-lief=0.12.3=py39h6a678d5_0
pycosat=0.6.4=py39h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py39h06a4308_0
pyparsing=3.0.9=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
pytest=7.3.1=py39h06a4308_0
pytest-cov=4.0.0=py39h06a4308_0
python=3.9.16=h955ad1f_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-libarchive-c=2.9=pyhd3eb1b0_1
python_abi=3.9=2_cp39
pytz=2022.7=py39h06a4308_0
pyyaml=6.0=py39h5eee18b_1
readline=8.2=h5eee18b_0
requests=2.29.0=py39h06a4308_0
ruamel.yaml=0.17.21=py39h5eee18b_0
ruamel.yaml.clib=0.2.6=py39h5eee18b_1
scipy=1.10.1=py39hf6e8229_1
setuptools=67.8.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.4=py39h06a4308_0
sqlalchemy=1.4.39=py39h5eee18b_0
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py39h06a4308_0
toolz=0.12.0=py39h06a4308_0
tqdm=4.65.0=py39hb070fc8_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py39h06a4308_0
wheel=0.38.4=py39h06a4308_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.11.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_0
zstandard=0.19.0=py39h5eee18b_0
zstd=1.5.5=hc292b87_0
